WIP on add-infinite-canvas
