+++ package.json
@@ -36,6 +36,7 @@
     "vue": "3.5.16",
     "vue-router": "4.5.1",
     "vue-sonner": "^1.3.2",
+    "vue-virtual-scroller": "^1.1.2",
     "zod": "^3.25.46"
   },
   "devDependencies": {
